+++ demo/FakeItEasy/Tethos.Xunit.Demo/ContainerFromBaseClass.cs
@@ -1,9 +1,9 @@
-namespace Tethos.xUnit.Demo
+namespace Tethos.Xunit.Demo
 {
-    using global::Moq;
-    using Tethos.Moq;
+    using global::FakeItEasy;
+    using global::Xunit;
+    using Tethos.FakeItEasy;
     using Tethos.Tests.Common;
-    using Xunit;
 
     public class ContainerFromBaseClass : AutoMockingTest
     {
@@ -14,10 +14,9 @@ namespace Tethos.xUnit.Demo
             // Arrange
             var expected = 42;
             var sut = this.Container.Resolve<SystemUnderTest>();
+            var mock = this.Container.Resolve<IMockable>();
 
-            this.Container.Resolve<Mock<IMockable>>()
-                .Setup(x => x.Do())
-                .Returns(expected);
+            A.CallTo(() => mock.Do()).Returns(expected);
 
             // Act
             var actual = sut.Do();
