+++ demo/FakeItEasy/Tethos.Xunit.Demo/ContainerInjected.cs
@@ -1,19 +1,19 @@
-﻿namespace Tethos.xUnit.Demo
+﻿namespace Tethos.Xunit.Demo
 {
     using System;
-    using global::Moq;
-    using Tethos.Moq;
+    using global::FakeItEasy;
+    using global::Xunit;
+    using Tethos.FakeItEasy;
     using Tethos.Tests.Common;
-    using Xunit;
 
-    public class ContainerAsProperty : IDisposable
+    public class ContainerInjected : IDisposable
     {
-        public ContainerAsProperty()
+        public ContainerInjected(IAutoFakeItEasyContainer container)
         {
-            this.Container = AutoMoqContainerFactory.Create();
+            this.Container = container;
         }
 
-        public IAutoMoqContainer Container { get; }
+        public IAutoFakeItEasyContainer Container { get; }
 
         [Fact]
         [Trait("", "Demo")]
@@ -22,10 +22,9 @@
             // Arrange
             var expected = 42;
             var sut = this.Container.Resolve<SystemUnderTest>();
+            var mock = this.Container.Resolve<IMockable>();
 
-            this.Container.Resolve<Mock<IMockable>>()
-                .Setup(x => x.Do())
-                .Returns(expected);
+            A.CallTo(() => mock.Do()).Returns(expected);
 
             // Act
             var actual = sut.Do();
