+++ demo/Moq/Tethos.Xunit.Demo/ContainerInjected.cs
@@ -1,19 +1,19 @@
-﻿namespace Tethos.xUnit.Demo
+﻿namespace Tethos.Xunit.Demo
 {
     using System;
-    using global::NSubstitute;
-    using Tethos.NSubstitute;
+    using global::Moq;
+    using global::Xunit;
+    using Tethos.Moq;
     using Tethos.Tests.Common;
-    using Xunit;
 
-    public class ContainerAsProperty : IDisposable
+    public class ContainerInjected : IDisposable
     {
-        public ContainerAsProperty()
+        public ContainerInjected(IAutoMoqContainer container)
         {
-            this.Container = AutoNSubstituteContainerFactory.Create();
+            this.Container = container;
         }
 
-        public IAutoNSubstituteContainer Container { get; }
+        public IAutoMoqContainer Container { get; }
 
         [Fact]
         [Trait("", "Demo")]
@@ -22,9 +22,10 @@
             // Arrange
             var expected = 42;
             var sut = this.Container.Resolve<SystemUnderTest>();
-            var mock = this.Container.Resolve<IMockable>();
 
-            mock.Do().Returns(expected);
+            this.Container.Resolve<Mock<IMockable>>()
+                .Setup(x => x.Do())
+                .Returns(expected);
 
             // Act
             var actual = sut.Do();
