+++ demo/NSubstitute/Tethos.Xunit.Demo/ContainerAsProperty.cs
@@ -1,12 +1,20 @@
-namespace Tethos.xUnit.Demo
+﻿namespace Tethos.Xunit.Demo
 {
+    using System;
     using global::NSubstitute;
+    using global::Xunit;
     using Tethos.NSubstitute;
     using Tethos.Tests.Common;
-    using Xunit;
 
-    public class ContainerFromBaseClass : AutoMockingTest
+    public class ContainerAsProperty : IDisposable
     {
+        public ContainerAsProperty()
+        {
+            this.Container = AutoNSubstituteContainerFactory.Create();
+        }
+
+        public IAutoNSubstituteContainer Container { get; }
+
         [Fact]
         [Trait("", "Demo")]
         public void Do_WithMock_ShouldReturn42()
@@ -23,6 +31,17 @@ namespace Tethos.xUnit.Demo
 
             // Assert
             Assert.Equal(actual, expected);
+        }
+
+        public void Dispose()
+        {
+            this.Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            this.Container?.Dispose();
         }
     }
 }
